requirements.txt
Flask==2.0.1
Pillow==8.4.0
opencv-python==4.5.3.56
opencv-contrib-python==4.8.1.78
numpy==1.24.3